+++ src/printer.js
@@ -1,5 +1,6 @@
 import * as doc from "prettier/doc";
 import embed from "./embed.js";
+import sortAttributesByKey from "./util/sortAttributesByKey.js";
 
 const { fill, group, hardline, indent, join, line, literalline, softline } =
   doc.builders;
@@ -369,9 +370,28 @@ function printElement(path, opts, print) {
   const parts = [OPEN[0].image, Name[0].image];
 
   if (attribute) {
+    const attributes = path.map(
+      (attributePath) => ({
+        node: attributePath.getValue(),
+        printed: print(attributePath)
+      }),
+      "children",
+      "attribute"
+    );
+
+    if (opts.xmlSortAttributesByKey) {
+      sortAttributesByKey(attributes);
+    }
+
     const separator = opts.singleAttributePerLine ? hardline : line;
     parts.push(
-      indent([line, join(separator, path.map(print, "children", "attribute"))])
+      indent([
+        line,
+        join(
+          separator,
+          attributes.map(({ printed }) => printed)
+        )
+      ])
     );
   }